+++ Docs/CHANGELOG.md
@@ -11,8 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated license view in OpenTgResearcherDesktop
 ### Added
 - Added methods to memorize window size and position between sessions in OpenTgResearcherDesktop
-- Added license request to TgDownloaderConsole
-- Added license request to OpenTgResearcherDesktop
+- Added license request
+- Added license clear
 - Increased the limit on the number of download threads for a user with a paid or test license
 
 ## [0.6.140] - 2025-04-26